# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Simple Gin API",
    "title": "Documentation of our Books API",
    "contact": {
      "name": "Test User",
      "url": "http://github.com/",
      "email": "some_email@example.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/books": {
      "get": {
        "description": "Returns list of books",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "books"
        ],
        "operationId": "listBooks",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        }
      },
      "post": {
        "description": "Adds a new book",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "books"
        ],
        "operationId": "addBook",
        "parameters": [
          {
            "description": "The new book to create",
            "name": "Book",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          {
            "type": "string",
            "example": "someJWTToken",
            "description": "Auth header, where the JWT token should be provided",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/books/search": {
      "get": {
        "description": "Filters list of books by tag",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "books"
        ],
        "operationId": "searchBooks",
        "parameters": [
          {
            "type": "string",
            "description": "tag to filter on",
            "name": "tag",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "description": "Returns a book",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "books"
        ],
        "operationId": "getBook",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the book",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "Update an existing book",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "books"
        ],
        "operationId": "updateBook",
        "parameters": [
          {
            "description": "The new book to create",
            "name": "Book",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          {
            "type": "string",
            "example": "someJWTToken",
            "description": "Auth header, where the JWT token should be provided",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Book"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "book not found"
          },
          "500": {
            "description": "internal server error"
          }
        }
      },
      "delete": {
        "description": "Delete an existing book",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "books"
        ],
        "operationId": "deleteBook",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the book",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "example": "someJWTToken",
            "description": "Auth header, where the JWT token should be provided",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/signin": {
      "post": {
        "description": "Signs in a user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signUser",
        "parameters": [
          {
            "description": "The user credentials",
            "name": "User",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "unauthorised"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/signout": {
      "post": {
        "description": "Signs out a user",
        "tags": [
          "auth"
        ],
        "operationId": "signUser",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "unauthorised"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "description": "Book represents a book",
      "type": "object",
      "required": [
        "name",
        "author",
        "publisher",
        "published_at",
        "review"
      ],
      "properties": {
        "author": {
          "description": "the book author's name",
          "type": "string",
          "minLength": 3,
          "x-go-name": "Author",
          "example": "Andrew Smith"
        },
        "image_path": {
          "description": "the image path for this book cover",
          "type": "string",
          "x-go-name": "ImagePath",
          "example": "\"/assets/images/book_cover.jpg\""
        },
        "name": {
          "description": "the name of the book",
          "type": "string",
          "minLength": 3,
          "uniqueItems": true,
          "x-go-name": "Name",
          "example": "Moondust"
        },
        "published_at": {
          "$ref": "#/definitions/PublishedDate"
        },
        "publisher": {
          "description": "The publisher's name",
          "type": "string",
          "minLength": 3,
          "x-go-name": "Publisher",
          "example": "Bloomsbury Publishing PLC"
        },
        "review": {
          "description": "Score review",
          "type": "number",
          "format": "float",
          "maximum": 5,
          "minimum": 0,
          "x-go-name": "Review",
          "example": 4.6
        },
        "submission_date": {
          "description": "The day the book was added",
          "type": "string",
          "format": "date-time",
          "x-go-name": "SubmissionDate"
        },
        "tags": {
          "description": "the associated tags with this book",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Tags",
          "example": [
            "space exploration",
            "astronauts",
            "nasa"
          ]
        }
      },
      "x-go-package": "github.com/PauloPortugal/gin-gonic-rest-mongodb/model"
    },
    "JWTOutput": {
      "type": "object",
      "title": "JWTOutput Represents the response after a successful attempt.",
      "required": [
        "token",
        "expires"
      ],
      "properties": {
        "expires": {
          "description": "Token's time-to-live (TTL)",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Expires"
        },
        "token": {
          "description": "the JWT token",
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "github.com/PauloPortugal/gin-gonic-rest-mongodb/model"
    },
    "PublishedDate": {
      "description": "PublishedDate Represents the month and year of when the book was published",
      "type": "object",
      "required": [
        "month",
        "year"
      ],
      "properties": {
        "month": {
          "description": "the month of the year",
          "type": "string",
          "x-go-name": "Month",
          "example": "July"
        },
        "year": {
          "description": "the month of the year",
          "type": "string",
          "minLength": 1,
          "x-go-name": "Year",
          "example": "2009"
        }
      },
      "x-go-package": "github.com/PauloPortugal/gin-gonic-rest-mongodb/model"
    },
    "User": {
      "type": "object",
      "title": "User represents the credentials of an actor trying to log in.",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "password": {
          "description": "the password",
          "type": "string",
          "x-go-name": "Password",
          "example": "password"
        },
        "username": {
          "description": "the username",
          "type": "string",
          "x-go-name": "Username",
          "example": "admin"
        }
      },
      "x-go-package": "github.com/PauloPortugal/gin-gonic-rest-mongodb/model"
    }
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}